[SW_EMU_TESTED] Poisson2d codegen u280 batching added

+++ codegen_apps/poisson2d/u280_loopback_project/user_types.h
@@ -6,6 +6,6 @@
 #define OPS_FUN_PREFIX
 #endif
 
-static inline OPS_FUN_PREFIX double myfun(double a, double b) {
+static inline OPS_FUN_PREFIX float myfun(float a, float b) {
   return a*b+1.0;
 }
